refactor: load the add AI provider form on demand
